track and display fouls and warnings
This feature can be turned on and off in app settings.
The changes to the main page include the addition of an "Add warning" and "Add foul" button, each of which take the user to a page were color, number, infraction, and equal puck for fouls or team warning. The infraction selection button is now also on the create penalty page.
Also on the main page, there is now a button that displays the three most recent warnings issued to each team. If this button is pressed, it will bring the user to a summary page of all warnings and fouls that were recorded in the game. Further work is needed (by Tristan) to have this page take the user to foul or warning edit pages.